Test: Add a memory access test case for isolation level 3
To verify if the boundary is set up correctly inside SPE, create a test
case to try to access memory in another ARoT partition. This is a
negative test and if tfm_hal_memory_has_access returns an error value,
the test succeeds. It is only enabled when isolation level is equal to 3.

Change-Id: Ie1d17c18d78ff24133ff3e5dede37db1c37b23e6
Signed-off-by: Shawn Shan <Shawn.Shan@arm.com>