Test: Support image update with dependency in FWU test suites
After image update with dependencies is supported in the Firmware
Update(FWU) partition, the FWU test suites should be modified
accordingly.

Added two test suites to test the image updates with dependency.

Also as after adding the image dependency check, the image data
used in the tests, as well as some test logic is coupled with
MCUboot. So the moved the FWU test suites into a 'mcuboot' folder.

Signed-off-by: Sherry Zhang <sherry.zhang2@arm.com>
Change-Id: I4c187fe816963adc8f4b6686150bdf0eaa0855d2