Build: Control single test without TEST_S/TEST_NS [Test repo]
To decrease the size of TF-M image so that it can work on a platform
with limited resource, the build configure should allow developers
to select one or more single test suites. The tf-m-test repo has
added some build control flags. These flags can be controlled with
"-D<FLAG>=ON" to turn on the related test suites' build. With the
build control flags are added in tf-m-tests repo, the cmake build
logic need to be refined. To make it easier to start single test
suite without TEST_NS or TEST_S, more conditions need to be added to
link tf-m-tests repo. Moreover, some test services should also
depend on the new single test flags.

By this patch, when TEST_S or TEST_NS is ON, all the secure or
non-secure test suites will run either. When single test configuration
like TEST_NS_XXX or TEST_S_XXX is ON, the single test suite will run
if its dependencies are configurated.

Signed-off-by: Jianliang Shen <jianliang.shen@arm.com>
Change-Id: I8351314d225e12f20859248d1868fa4bc6cfc26b