build: suites: exclude attestation if not required
Restricts including the `attestation`, `qcbor`, and `t_cose` folders to
situations where initial attestation services are actually required
(`TEST_S_ATTESTATION` set). This check is required to avoid downloading
QCBOR at compile time, which is pulled in as a dependency of t_cose.

This patch should be reverted once an acceptable upstream solution
can be found.

Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>