RPC: Remove option of connection-based API
The connection-based APIs are now always implemented on NS client side
(eRPC Server).
It is not necessary to optionally build them because the config option
is set manually anyway.

This patch removes the optional build and makes them built always.
The grouping of the APIs are kept in case of future use.

Change-Id: I798894b3e8f97f0bff336e0db185ee744ca70f87
Signed-off-by: Kevin Peng <kevin.peng@arm.com>