Build: Merge tfm_ns_interface to tfm_api_ns
The tfm_ns_interface is an INTERFACE lib that collects include
directories and compile definitions for NS building.

The tfm_api_ns mainly contains the source files.

They are basically the same thing seperated to two parts.
This is not necessary from the CMake point of view.

Merging them into one and make the include directoris and compile
definitions PUBLIC should simply do.

Change-Id: I5a1c020dc5fb07c0451e7be3218522a932ccc952
Signed-off-by: Kevin Peng <kevin.peng@arm.com>