Build: Remove some common SPM test components
1. The spm_test_service_common library

For SPE build, it is required by test Secure Partitions and it's
built only when Secure test suites are enabled.

There are some cases when a test partition is enabled but no Secure
tests are enabled and then this library wouldn't be built.

This patch removes it. It only provides include directories so add them
in the libraries that link to it directly.

2. The common tfm_irq_test_service.c

The functions within it are only referenced by individual files which
means they are not "common".

This patch moves the functions into those files and removes this file.

Change-Id: I85a1d2cd04b4261da8dd85c200c05f5abf68a0f5
Signed-off-by: Kevin Peng <kevin.peng@arm.com>