Test: Add USE_STDIO option for test logs
The clients like RPC test framework hosts may use the stdio provided by
toolchain.
This patch adds this config option for test log.
It also removes the unncessary inclusions of <stdio.h>.
They are partially for the log function and the rest for the <stddef.h>
which is included by stdio.h.

Change-Id: I79328d50b43e6c0b0d7ed55296c9973a90cb90e1
Signed-off-by: Kevin Peng <kevin.peng@arm.com>