Test Framework: Rename integ_test() and move it
The fucntion name of integ_test() does not reflect what it does.
It just runs the test suites given, not particualr to "integ test" only.

Rename it to run_test() and move it to test_framework.h/.c.
It does not need to be put in a dedicated set of files.

Change-Id: I264ceb4c7b48b74e57694c8db688a3da780a21d1
Signed-off-by: Kevin Peng <kevin.peng@arm.com>