Build: Change NS build to be one of mailbox or TZ
Refactors the NS build so that it is either built for mailbox or for
TrustZone, but not both. TF-M allows platforms to have both TZ and
mailbox agents, but the NS build must be done for the local and remote
cores separately. Having both the mailbox and TZ APIs built into the
same binary results in link errors due to multiple definition of the
PSA functions.

The TFM_NS_MAILBOX_API flag is set ON by default if TF-M was built with
the mailbox agent and not the TZ agent, OFF by default otherwise. The
default only needs to be overridden to do a mailbox build on platforms
with both agents.

Change-Id: Ic296dd7cb4141ff5ec5f008962c9fe9459ecff5f
Signed-off-by: Jamie Fox <jamie.fox@arm.com>