Crypto: Restructure dependencies on psa_crypto_config
Targets that are explicit callers of PSA Crypto APIs (i.e. that
need to include psa/crypto.h) need to be explicitly linked with
psa_crypto_config. Also, the psa_crypto_config now is a proper
target that gets exported by the SPE build, so make sure it is
visible to psa_arch_tests val_nspe/pal_nspe and that in turn
links against the platform so that platform includes are available


Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: Ib75d090397ef15df26fd0c3677141215855d133c