eRPC: add platform_region_defs to host build
Add the platform_region_defs interface target to the host (eRPC client)
image build as other targets may depend on the region definitions or
flash layout such as the FWU test suite.
During a regular split build it's provided by the spe-CMakeLists.cmake
file but with the eRPC framework it was missing from the host image build.

Change-Id: Ia7b3583ae402a2f01b35405951d40803739d7fc0
Signed-off-by: David Vincze <david.vincze@arm.com>