FPU: Avoid linking to test framework from partition
The FPU test partition was linking to the TF-M test framework to access
shared FPU functions in the tests. However, as the FPU partition is App
RoT, it cannot call these functions when stricter code isolation is
applied. To solve the issue, this patch splits the shared FPU library
functions out from the common test code and adds them to TF-M SPRT,
which can be called by any Secure code.

Change-Id: If46249201c5ff8c4a30dc12a80d731e649780961
Signed-off-by: Jamie Fox <jamie.fox@arm.com>