attest: fix check for ATTEST_TOKEN_PROFILE_ARM_CCA
The macro may be defined and set to 0 to indicate that this profile
is not enabled, in which case the check previously considered that
the profile was enabled while it was in fact not, resulting in
macro redifinitions with different values.

Change the check from #ifdef to #if to fix that.

Change-Id: Ib5f95da2d877e6ef4f72a472bbbdf2716b8291ee
Signed-off-by: Tomi Fontanilles <tomi.fontanilles@nordicsemi.no>
(cherry picked from commit 495042b84a13f883a3361f9f9aeb82704d1a82d9)