framework: Avoid compiler error in test_err_to_str
The compiler need not assume that the only valid values
for err are those defined in the enum, so it may think
that the function falls off the end. Add an assert(0)
to cause the test to fail if that actually happens.

The compiler can still validate that all enum values are present in
the switch statement, providing the desired diagnostic at compile
time.

Signed-off-by: Keith Packard <keithp@keithp.com>